Implement the logic for disable the extend option if it was used previously during current session and reset after the session complete and stop. Implement the logic for adding the unblock list and block list of General setting and Category list from Edit block list. (Continue)

+++ Focus/Controller/Setting/General/GeneralSettingViewC.swift
@@ -163,23 +163,22 @@ extension GeneralSettingViewC: BasicSetupType {
         popBreakTime.menu = Focus.BreakTime.breaktimes
         popBreakTime.target = self
         popBreakTime.action = #selector(breakTimeSelection(_:))
-        
-        disableControl()
     }
 
-    func disableControl() {
-        guard let objF = DBManager.shared.getCurrentBlockList().objFocus else { return }
+    func disableControl() -> Bool {
+        guard let objF = DBManager.shared.getCurrentBlockList().objFocus else { return false }
 
-        btnAddWeb.isEnabled = !objF.is_focusing
-        btnAddApp.isEnabled = !objF.is_focusing
-        tblView.isEnabled = !objF.is_focusing
+        if objF.is_focusing {
+            openErrorDialogue()
+            setupData()
+        }
+        return true
+    }
 
-        checkBoxWarning.isEnabled = !objF.is_focusing
-        checkBoxEachBreak.isEnabled = !objF.is_focusing
-        checkBoxFocusTime.isEnabled = !objF.is_focusing
-        checkBoxShowTimer.isEnabled = !objF.is_focusing
-        popFocusTime.isEnabled = !objF.is_focusing
-        popBreakTime.isEnabled = !objF.is_focusing
+    func openErrorDialogue() {
+        let errorDialog = ErrorDialogueViewC(nibName: "ErrorDialogueViewC", bundle: nil)
+        errorDialog.errorType = .general_setting_error
+        presentAsSheet(errorDialog)
     }
 }
 
@@ -210,7 +209,9 @@ extension GeneralSettingViewC {
 
     @IBAction func checkBoxEventHandler(_ sender: NSButton) {
         print(" Check Box Event : \(sender.state) ::: \(sender.title)")
-
+        if disableControl() {
+            return
+        }
         let isChecked = (sender.state == .on) ? true : false
         switch sender.tag {
         case 0:
@@ -228,6 +229,10 @@ extension GeneralSettingViewC {
     }
 
     @objc func addAppAction(_ sender: NSButton) {
+        if disableControl() {
+            return
+        }
+
         let objCat = viewModel.input.getGeneralCategoryData().gCat
         if objCat != nil {
             let controller = BlocklistDialogueViewC(nibName: "BlocklistDialogueViewC", bundle: nil)
@@ -246,6 +251,10 @@ extension GeneralSettingViewC {
     }
 
     @objc func addWebAction(_ sender: NSButton) {
+        if disableControl() {
+            return
+        }
+
         let objCat = viewModel.input.getGeneralCategoryData().gCat
         if objCat != nil {
             let inputDialogueCntrl = InputDialogueViewC(nibName: "InputDialogueViewC", bundle: nil)
@@ -264,6 +273,10 @@ extension GeneralSettingViewC {
     }
 
     @objc func deleteSubCate(_ sender: NSButton) {
+        if disableControl() {
+            return
+        }
+
         let arrSCat = viewModel.objGCategory?.sub_data?.allObjects as? [Block_SubCategory]
         guard let objBlock = arrSCat?[sender.tag] else { return }
         DBManager.shared.managedContext.delete(objBlock)
@@ -273,6 +286,10 @@ extension GeneralSettingViewC {
 
     // Store the Categories as per selected blick list
     @objc func addSCategory(_ sender: NSButton) {
+        if disableControl() {
+            return
+        }
+
         let arrSCat = viewModel.objGCategory?.sub_data?.allObjects as? [Block_SubCategory]
         guard let objBlock = arrSCat?[sender.tag] else { return }
         objBlock.is_selected = !objBlock.is_selected
